Follow symlinks when creating directories
It looks like the directory creation process now errors out when you want to create a directory that exists as a symlink to a directory.  It complains that it's a symlink.  `follow=yes` does the right thing in that it'll not try to create the directory if the symlink exists and it points to a directory.